requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
email-validator
mysql-connector-python
